requirements.txt
absl-py==2.3.1
accelerate==1.4.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.4.0
boto3==1.40.63
botocore==1.40.63
cachetools==6.2.1
certifi==2025.10.5
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.2
cloudpickle==3.1.1
codetiming==1.4.0
colorama==0.4.6
colorful==0.5.8
contourpy==1.3.2
cpm-kernels==1.0.11
cycler==0.12.1
DataProperty==1.1.0
datasets==4.0.0
dill==0.3.8
distlib==0.4.0
distro==1.9.0
einops==0.8.1
evaluate==0.4.6
exceptiongroup==1.3.0
fastapi==0.120.2
filelock==3.20.0
fire==0.7.1
flash_attn==2.7.4.post1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
func_timeout==4.3.5
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.42.0
googleapis-common-protos==1.71.0
gradio_client==1.13.3
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.36.0
hydra-core==1.3.2
identify==2.6.15
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2sympy2_extended==1.10.2
Levenshtein==0.27.3
liger_kernel==0.6.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
math-verify==0.8.0
matplotlib==3.10.7
mbstrdecoder==1.1.4
mdurl==0.1.2
mmengine-lite==0.10.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.2
nodeenv==1.9.1
numexpr==2.14.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==2.6.1
OpenCC==1.1.9
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.38.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
pandas==1.5.3
pathvalidate==3.3.1
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
pre_commit==4.3.0
prettytable==3.16.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.2
py-spy==0.4.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pydantic==2.12.3
pydantic_core==2.41.4
pyecharts==2.0.9
pyext==0.7
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
ray==2.51.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
retrying==1.4.2
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
sacrebleu==2.5.1
safetensors==0.6.2
scikit-learn==1.5.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.1.2
sentry-sdk==2.43.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smart_open==7.4.2
smmap==5.0.2
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.49.1
swanlab==0.6.13
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tempdir==0.7.1
tensordict==0.9.0
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timeout-decorator==0.5.0
tokenizers==0.21.4
tomli==2.3.0
torch==2.6.0
torchdata==0.11.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.52.4
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
triton==3.2.0
typepy==1.3.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
wandb==0.22.3
wcwidth==0.2.14
websockets==15.0.1
wget==3.2
word2number==1.1
wrapt==2.0.0
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
